move get_catalog_info() to common.py
