+++ flowsa/metadata.py
@@ -9,7 +9,7 @@ FlowByActivity (FBA) and FlowBySector (FBS) datasets
 import pandas as pd
 from esupy.processed_data_mgmt import FileMeta, write_metadata_to_file, \
     read_source_metadata
-from flowsa.common import return_true_source_catalog_name
+from flowsa.common import return_true_source_catalog_name, get_catalog_info
 from flowsa.flowsa_log import log
 from flowsa.settings import paths, PKG, PKG_VERSION_NUMBER, WRITE_FORMAT, \
     GIT_HASH, GIT_HASH_LONG
@@ -94,7 +94,6 @@ def return_fbs_method_data(source_name, config):
     """
     from flowsa.data_source_scripts.stewiFBS import add_stewi_metadata,\
         add_stewicombo_metadata
-    from flowsa.flowby import get_catalog_info
 
     # Create empty dictionary for storing meta data
     meta = {}
